add function to identify sudden losses.
or sudden gains on scales where higher values are "better") here however the second criterion would work differently as 25percent of a high value are different to 25%  of a low value and therefore using the sl function to identify sg would be less strict. however this shouldnt be too much of a problem as the second crit doesnt really deviate much from first crit, this gets sometimes discussed in the literature.